split project and scene setup windows
added framework for help window
reordered cognitive3d menu
added placeholder layout for dynamic object window
removed scene upload buttons from preferences
