+++ Editor/HelpWindow.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 8d63bbc04509108429a87ad11794fa6c
+timeCreated: 1520470184
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
