+++ Editor/SceneSetupWindow.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 14b41bdf5a28d1e46b987922faecd7a8
+timeCreated: 1520470184
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
