+++ package.json
@@ -8,7 +8,8 @@
 	"changelogUrl": "https://github.com/CognitiveVR/cvr-sdk-unity/releases/",
 	"licensesUrl": "https://github.com/CognitiveVR/cvr-sdk-unity/blob/master/LICENSE",
 	"dependencies": {
-		"com.unity.xr.legacyinputhelpers": "2.1.7"		
+		"com.unity.xr.legacyinputhelpers": "2.1.7",
+		"com.unity.xr.management":"4.2.1"
 	},
 	"keywords": [
 		"xr",
